+++ client/app/reader/PdfViewer.jsx
@@ -9,7 +9,7 @@ import Modal from '../components/Modal';
 import { closeAnnotationDeleteModal, deleteAnnotation,
   handleSelectCommentIcon, selectCurrentPdf } from '../reader/actions';
 import { bindActionCreators } from 'redux';
-import { getAnnotationByDocumentId } from '../reader/util/AnnotationUtil';
+import { getAnnotationByDocumentId } from '../reader/utils';
 
 // PdfViewer is a smart component that renders the entire
 // PDF view of the Reader SPA. It displays the PDF with UI
@@ -165,7 +165,7 @@ export class PdfViewer extends React.Component {
 }
 
 const mapStateToProps = (state, ownProps) => ({
-  editingCommentsForCurrentDoc: 
+  editingCommentsForCurrentDoc:
     _.some(getAnnotationByDocumentId(state, Number(ownProps.match.params.docId)), 'editing'),
   ..._.pick(state.ui, 'deleteAnnotationModalIsOpenFor', 'placedButUnsavedAnnotation'),
   ..._.pick(state.ui.pdf, 'commentFlowState', 'scrollToComment', 'hidePdfSidebar')
